vpn_backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
web3==5.24.0
pyjwt==2.3.0
cryptography==3.4.7
paramiko==2.8.1  # For SSH connections to VPN nodes 